Separate initialization to allow for testing without starting server
